+++ src/registrar/config/urls.py
@@ -25,7 +25,7 @@ from registrar.views.domain_request import Step
 from registrar.views.domain_requests_json import get_domain_requests_json
 from registrar.views.domains_json import get_domains_json
 from registrar.views.utility import always_404
-from registrar.views.index_organizations import index_organizations
+from registrar.views.organizations import index_organizations, organization_domains, organization_domain_requests
 from api.views import available, get_current_federal, get_current_full
 
 
@@ -63,6 +63,16 @@ urlpatterns = [
         "organization/<int:portfolio_id>/",
         index_organizations,
         name="home-organization",
+    ),
+    path(
+        "organization/<int:portfolio_id>/domains/",
+        organization_domains,
+        name="organization-domains",
+    ),
+    path(
+        "organization/<int:portfolio_id>/domain_requests/",
+        organization_domain_requests,
+        name="organization-domain-requests",
     ),
     path(
         "admin/logout/",
