+++ src/registrar/templates/organization_sidebar.html
@@ -4,13 +4,16 @@
   <nav aria-label="">
     <ul class="usa-sidenav">
       <li class="usa-sidenav__item">
-        <a href="#"> 
-           link 1
+        {{ portfolio.organization_name }}
+      </li>
+      <li class="usa-sidenav__item">
+        <a href="{% url 'organization-domains' portfolio.id %}" class="{% if current_path == 'domains' %}usa-current{% endif %}"> 
+           Domains
         </a>
       </li>
       <li class="usa-sidenav__item">
-        <a href="#"> 
-           link 2
+        <a href="{% url 'organization-domain-requests' portfolio.id %}" class="{% if current_path == 'domain-requests' %}usa-current{% endif %}"> 
+           Domain requests
         </a>
       </li>
     </ul>
